requirements.txt
python-dotenv
langchain~=0.0.225
gpt4all~=1.0.1
unstructured
pinecone-client
pydantic~=1.10.11
click~=8.1.4
torch~=2.0.1
numpy~=1.23.5
transformers~=4.30.2
sentencepiece~=0.1.99
typer~=0.7.0
requests~=2.31.0
tqdm~=4.65.0
setuptools~=68.0.0
PyYAML~=6.0
pandas~=1.5.3
scikit-learn~=1.3.0
pyarrow~=12.0.1
streamlit~=1.24.0